fix: retry RPC call to local replicator
