+++ jobs/src/main/scala/redsort/jobs/scheduler/WorkerRpcClientFiber.scala
@@ -20,6 +20,7 @@ import io.grpc.Status
 import redsort.jobs.messages.JobResult
 import org.scalatest.Assertion
 import redsort.jobs.messages.WorkerError
+import redsort.jobs.RPChelper
 
 object WorkerRpcClientFiber {
   private[this] val logger = new SourceLogger(getLogger, "scheduler")
@@ -96,30 +97,11 @@ object WorkerRpcClientFiber {
   ): IO[JobResult] = {
     rpcClient
       .runJob(spec, new Metadata)
-      .handleErrorWith(handleRpcErrorWithRetry(runJobWithRetry(rpcClient, spec)))
+      .handleErrorWith(RPChelper.handleRpcErrorWithRetry(runJobWithRetry(rpcClient, spec)))
   }
 
   private def completeWithRetry(rpcClient: WorkerFs2Grpc[IO, Metadata]): IO[Empty] =
     rpcClient
       .complete(new Empty, new Metadata)
-      .handleErrorWith(handleRpcErrorWithRetry(completeWithRetry(rpcClient)))
-
-  // NOTE: lazy evaluation (=>) is required - otherwise stack will overflow
-  private def handleRpcErrorWithRetry[T](retry: => IO[T])(err: Throwable): IO[T] =
-    err match {
-      case err if isTransportError(err) =>
-        for {
-          _ <- logger.error(s"transport error: $err")
-          _ <- logger.error(s"RPC call failed due to transport error, will retry in 1 seconds...")
-          _ <- IO.sleep(1.second)
-          result <- retry
-        } yield result
-      case err =>
-        logger.error(s"RPC call raised fatal exception: $err") >> IO.raiseError[T](err)
-    }
-
-  private def isTransportError(err: Throwable): Boolean = err match {
-    case e: StatusRuntimeException => e.getStatus.getCode == Status.Code.UNAVAILABLE
-    case _                         => false
-  }
+      .handleErrorWith(RPChelper.handleRpcErrorWithRetry(completeWithRetry(rpcClient)))
 }
